backend_BI/app/requirement.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tabulate==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.2
